requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
chardet==5.2.0
dnspython==2.6.1
frozenlist==1.4.1
idna==3.8
magic-filter==1.0.12
motor==3.5.1
multidict==6.0.5
numpy==2.1.1
pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pymongo==4.8.0
python-dotenv==1.0.1
reportlab==4.2.2
typing_extensions==4.12.2
yarl==1.11.0
